[#23] feat: S3 이미지 upload 구현
