+++ b2b/src/main/java/com/sparta/b2b/fileUpload/dto/ImageInfo.java
@@ -1,0 +1,14 @@
+package com.sparta.b2b.fileUpload.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public  class ImageInfo {
+	private String url;
+}
