+++ b2b/src/main/java/com/sparta/b2b/product/controller/ProductController.java
@@ -28,7 +28,7 @@ public class ProductController {
 	public ResponseEntity<ProductCreateResponse> createProduct(
 		@RequestBody @Valid ProductCreateRequest request,
 		@LoginMember(role = Role.B2B) MemberSession memberSession
-		
+
 	) {
 		return ResponseEntity
 			.status(HttpStatus.CREATED)
