+++ domain/src/main/java/com/sparta/impostor/commerce/backend/common/config/S3Config.java
@@ -16,7 +16,7 @@ public class S3Config {
 	private String accessKey;
 	@Value("${cloud.aws.credentials.secretKey}")
 	private String secretKey;
-	@Value("${cloud.aws.region.static}")
+	@Value("${cloud.aws.s3.region}")
 	private String region;
 
 	@Bean
